@@ -57,7 +57,8 @@ export const CategoryController: IController = {
         res.status(403).send({
           error: "You are not authorized to create a category",
         });
-        await unlink(`./public/category/${req.file?.filename}`);
+        if (req.file !== undefined)
+          await unlink(`./public/category/${req.file.filename}`);
       }
 
       // check if name is string or not empty
@@ -66,21 +67,22 @@ export const CategoryController: IController = {
       ): Promise<void> => {
         if (validator.isEmpty(value, { ignore_whitespace: true })) {
           res.status(422).send({ error: `Please fill the empty field` });
-          await unlink(`./public/category/${req.file?.filename}`);
+          if (req.file !== undefined)
+            await unlink(`./public/category/${req.file?.filename}`);
           return;
         }
         if (typeof value !== "string") {
           res
             .status(400)
             .send({ error: `Field must contains only characters` });
-          await unlink(`./public/category/${req.file?.filename}`);
-
+          if (req.file !== undefined)
+            await unlink(`./public/category/${req.file?.filename}`);
           return;
         }
       };
 
       const inputs: string[] = Object.values(req.body);
-      inputs.forEach((value) => checkIfEmptyAndNotAString(value));
+      inputs.forEach(async (value) => await checkIfEmptyAndNotAString(value));
 
       // Check format of category's name
       if (
@@ -92,16 +94,18 @@ export const CategoryController: IController = {
         res.status(400).send({
           error: `Field must contains only characters (min: 2, max: 100)`,
         });
-        await unlink(`./public/category/${req.file?.filename}`);
+        if (req.file !== undefined)
+          await unlink(`./public/category/${req.file?.filename}`);
         return;
       }
 
       // check if image is an object
-      if (image && typeof image !== "object") {
+      if (image !== null && typeof image !== "object") {
         res.status(400).send({
           error: `Field image must contains a file`,
         });
-        await unlink(`./public/category/${req.file?.filename}`);
+        if (req.file !== undefined)
+          await unlink(`./public/category/${req.file?.filename}`);
         return;
       }
 
@@ -113,12 +117,13 @@ export const CategoryController: IController = {
         res.status(409).send({
           error: `Category already exists`,
         });
-        await unlink(`./public/category/${req.file?.filename}`);
+        if (req.file !== undefined)
+          await unlink(`./public/category/${req.file?.filename}`);
         return;
       }
 
       // rename file image
-      if (req.file) {
+      if (req.file !== undefined) {
         const originalname = req.file.originalname;
         const filename = req.file.filename;
         const newName = `${uuidv4()}-${originalname}`;
@@ -126,7 +131,7 @@ export const CategoryController: IController = {
           `./public/category/${filename}`,
           `./public/category/${newName}`,
           (err) => {
-            if (err) throw err;
+            if (err !== null) throw err;
           }
         );
         req.body.image = `/public/category/${newName}`;
@@ -139,7 +144,8 @@ export const CategoryController: IController = {
       res.status(201).send("Created category");
     } catch (error) {
       res.status(400).send({ error: "Something went wrong" });
-      await unlink(`./public/category/${req.file?.filename}`);
+      if (req.file !== undefined)
+        await unlink(`./public/category/${req.file?.filename}`);
     }
   },
 
@@ -161,7 +167,8 @@ export const CategoryController: IController = {
         res.status(403).send({
           error: "You are not authorized to create a category",
         });
-        await unlink(`./public/category/${req.file?.filename}`);
+        if (req.file !== undefined)
+          await unlink(`./public/category/${req.file?.filename}`);
       }
 
       // validate format
@@ -176,19 +183,20 @@ export const CategoryController: IController = {
           res
             .status(400)
             .send({ error: `Field must contains only characters` });
-          await unlink(`./public/category/${req.file?.filename}`);
+          if (req.file !== undefined)
+            await unlink(`./public/category/${req.file?.filename}`);
         }
       };
 
       // check if values are string and not empty
 
       const inputs: string[] = Object.values(req.body);
-      inputs.forEach((value) => checkIfEmptyAndNotAString(value));
+      inputs.forEach(async (value) => await checkIfEmptyAndNotAString(value));
 
       // check if name contains only characters
 
       if (
-        name &&
+        name !== null &&
         !validator.matches(
           name,
           /^[a-zA-ZàáâäãåąčćęèéêëėįìíîïłńòóôöõøùúûüųūÿýżźñçčšžÀÁÂÄÃÅĄĆČĖĘÈÉÊËÌÍÎÏĮŁŃÒÓÔÖÕØÙÚÛÜŲŪŸÝŻŹÑßÇŒÆČŠŽ∂ð '-]{2,100}$/
@@ -197,17 +205,19 @@ export const CategoryController: IController = {
         res.status(400).send({
           error: `Field must contains only characters (min: 2, max: 100)`,
         });
-        await unlink(`./public/category/${req.file?.filename}`);
+        if (req.file !== undefined)
+          await unlink(`./public/category/${req.file?.filename}`);
 
         return;
       }
 
       // check if image is an object
-      if (image && typeof image !== "object") {
+      if (image !== null && typeof image !== "object") {
         res.status(400).send({
           error: `Field image must contains a file`,
         });
-        await unlink(`./public/category/${req.file?.filename}`);
+        if (req.file !== undefined)
+          await unlink(`./public/category/${req.file?.filename}`);
 
         return;
       }
@@ -219,13 +229,14 @@ export const CategoryController: IController = {
 
       if (categoryToUpdate === null) {
         res.status(404).send({ error: "Category not found" });
-        await unlink(`./public/category/${req.file?.filename}`);
+        if (req.file !== undefined)
+          await unlink(`./public/category/${req.file?.filename}`);
 
         return;
       }
 
       // check if body.name already exists in db
-      if (name) {
+      if (name !== null) {
         const nameAlreadyExist = await dataSource
           .getRepository(Category)
           // check every tuples except the one updating
@@ -233,14 +244,15 @@ export const CategoryController: IController = {
 
         if (nameAlreadyExist > 0) {
           res.status(409).send({ error: "Category name already exist" });
-          await unlink(`./public/category/${req.file?.filename}`);
+          if (req.file !== undefined)
+            await unlink(`./public/category/${req.file?.filename}`);
 
           return;
         }
       }
 
       // rename the new file and delete the old one
-      if (req.file) {
+      if (req.file !== undefined) {
         // rename
         const originalname = req.file.originalname;
         const filename = req.file.filename;
@@ -249,12 +261,12 @@ export const CategoryController: IController = {
           `./public/category/${filename}`,
           `./public/category/${newName}`,
           (err) => {
-            if (err) throw err;
+            if (err !== null) throw err;
           }
         );
         req.body.image = `/public/category/${newName}`;
         // delete
-        if (categoryToUpdate.image) {
+        if (categoryToUpdate.image !== null) {
           await unlink("." + categoryToUpdate.image);
         }
       }
@@ -263,7 +275,8 @@ export const CategoryController: IController = {
       res.status(200).send("Updated category");
     } catch (err) {
       res.status(400).send({ error: "Error while updating category" });
-      await unlink(`./public/category/${req.file?.filename}`);
+      if (req.file !== undefined)
+        await unlink(`./public/category/${req.file?.filename}`);
     }
   },
 
