@@ -18,32 +18,66 @@ export const PoiController: IController = {
       let searchQueries = {};
 
       // if query in url, add finding options
-      if (req.query) {
+      if (Object.keys(req.query).length > 0) {
         const city = req.query.city as string;
         const category = req.query.category as string;
-        if (city) {
-          searchQueries = { city: { name: city } };
-        } else if (category) {
-          searchQueries = { category: { name: category } };
-        } else if (city && category) {
+        if (city !== undefined && category !== undefined) {
           searchQueries = {
             category: { name: category },
             city: { name: city },
           };
+        } else if (city !== undefined) {
+          searchQueries = { city: { name: city } };
+        } else if (category !== undefined) {
+          searchQueries = { category: { name: category } };
         }
       }
 
-      // get all poi
-      let allPoi = await dataSource.getRepository(Poi).find({
-        relations: {
-          category: true,
-          city: true,
-          user: true,
-        },
-        where: searchQueries,
-      });
+      // get poi is accepted or no depending of user status (admin or not)
+      const { userId } = req.params;
 
-      res.status(200).send(allPoi);
+      const currentUser = await dataSource
+        .getRepository(User)
+        .findOne({ where: { id: userId } });
+
+      let allPoi;
+      if (currentUser?.role === UserRole.ADMIN) {
+        // get all poi (accepted or not)
+        allPoi = await dataSource.getRepository(Poi).find({
+          relations: {
+            category: true,
+            city: true,
+            user: true,
+          },
+          where: searchQueries,
+        });
+      } else if (currentUser?.role === UserRole.ADMIN_CITY) {
+        // get the city name where user is admin to get only local POIs
+        const cityWhereUserIsAdmin = await dataSource
+          .getRepository(City)
+          .findOneBy({ userAdminCity: { id: userId } });
+        // update searchQueries with the city where user is admin
+        searchQueries = { ...searchQueries, city: cityWhereUserIsAdmin?.name };
+        allPoi = await dataSource.getRepository(Poi).find({
+          relations: {
+            category: true,
+            city: true,
+            user: true,
+          },
+          where: searchQueries,
+        });
+      } else {
+        // get only accepted poi
+        allPoi = await dataSource.getRepository(Poi).find({
+          relations: {
+            category: true,
+            city: true,
+            user: true,
+          },
+          where: { ...searchQueries, isAccepted: true },
+        });
+        res.status(200).send(allPoi);
+      }
     } catch (err) {
       res.status(400).send({ error: "Error while reading points of interest" });
     }
@@ -72,7 +106,7 @@ export const PoiController: IController = {
       const {
         description,
         address,
-        is_accepted,
+        isAccepted,
         name,
         image,
         category,
@@ -102,8 +136,8 @@ export const PoiController: IController = {
         res.status(403).send({
           error: "You are not authorized to create a point of interest",
         });
-        await unlink(`./public/poi/${req.file?.filename}`);
-
+        if (req.file !== undefined)
+          await unlink(`./public/poi/${req.file.filename}`);
         return;
       }
 
@@ -117,14 +151,15 @@ export const PoiController: IController = {
           res
             .status(400)
             .send({ error: `Field must contains only characters` });
-          await unlink(`./public/poi/${req.file?.filename}`);
-
-          return;
+          if (req.file !== undefined)
+            await unlink(`./public/poi/${req.file?.filename}`);
         }
       };
 
       const inputString: string[] = [description, name, address];
-      inputString.forEach((value) => checkIfStringAndNotEmpty(value));
+      inputString.forEach(
+        async (value) => await checkIfStringAndNotEmpty(value)
+      );
 
       // check name
 
@@ -137,8 +172,8 @@ export const PoiController: IController = {
         res.status(400).send({
           error: `Field must contains only characters (min: 2, max: 100)`,
         });
-        await unlink(`./public/poi/${req.file?.filename}`);
-
+        if (req.file !== undefined)
+          await unlink(`./public/poi/${req.file.filename}`);
         return;
       }
 
@@ -153,8 +188,8 @@ export const PoiController: IController = {
         res.status(400).send({
           error: `Incorrect format of address`,
         });
-        await unlink(`./public/poi/${req.file?.filename}`);
-
+        if (req.file !== undefined)
+          await unlink(`./public/poi/${req.file.filename}`);
         return;
       }
 
@@ -165,31 +200,32 @@ export const PoiController: IController = {
           res
             .status(400)
             .send({ error: "Incorrect format of foreign key (must be uuid)" });
-          await unlink(`./public/poi/${req.file?.filename}`);
-
-          return;
+          if (req.file !== undefined)
+            await unlink(`./public/poi/${req.file?.filename}`);
         }
       };
       const foreignKeys: string[] = [category, city, user];
-      foreignKeys.forEach((value) => {
-        if (value) checkIfUUID(value);
+      foreignKeys.forEach(async (value) => {
+        if (value !== undefined) await checkIfUUID(value);
       });
 
-      // check if is_accepted is boolean
+      // check if isAccepted is boolean
 
-      if (!validator.isBoolean(is_accepted)) {
+      if (!validator.isBoolean(isAccepted)) {
         res.status(400).send({ error: "is_accepted must be a boolean" });
-        await unlink(`./public/poi/${req.file?.filename}`);
+        if (req.file !== undefined)
+          await unlink(`./public/poi/${req.file.filename}`);
         return;
       }
 
       // check if image is an object
 
-      if (image && typeof image !== "object") {
+      if (image !== undefined && typeof image !== "object") {
         res.status(400).send({
           error: `Field image must contains a file`,
         });
-        await unlink(`./public/poi/${req.file?.filename}`);
+        if (req.file !== undefined)
+          await unlink(`./public/poi/${req.file.filename}`);
         return;
       }
 
@@ -198,7 +234,8 @@ export const PoiController: IController = {
         res.status(400).send({
           error: "Incorrect format of coordinates (must be [lat, long])",
         });
-        await unlink(`./public/poi/${req.file?.filename}`);
+        if (req.file !== undefined)
+          await unlink(`./public/poi/${req.file.filename}`);
         return;
       }
 
@@ -213,7 +250,8 @@ export const PoiController: IController = {
       });
       if (coordsAlreadyExist > 0) {
         res.status(409).send({ error: "Point of interest already exists" });
-        await unlink(`./public/poi/${req.file?.filename}`);
+        if (req.file !== undefined)
+          await unlink(`./public/poi/${req.file.filename}`);
         return;
       }
 
@@ -224,7 +262,7 @@ export const PoiController: IController = {
       };
 
       // rename file image
-      if (req.file) {
+      if (req.file !== undefined) {
         const originalname = req.file.originalname;
         const filename = req.file.filename;
         const newName = `${uuidv4()}-${originalname}`;
@@ -232,7 +270,7 @@ export const PoiController: IController = {
           `./public/poi/${filename}`,
           `./public/poi/${newName}`,
           (err) => {
-            if (err) throw err;
+            if (err !== null) throw err;
           }
         );
         req.body.image = `/public/poi/${newName}`;
@@ -245,7 +283,8 @@ export const PoiController: IController = {
       res.status(201).send("Created point of interest");
     } catch (err) {
       res.status(400).send({ error: "Something went wrong" });
-      await unlink(`./public/poi/${req.file?.filename}`);
+      if (req.file !== undefined)
+        await unlink(`./public/poi/${req.file.filename}`);
     }
   },
 
@@ -256,7 +295,7 @@ export const PoiController: IController = {
       const {
         description,
         address,
-        is_accepted,
+        isAccepted,
         name,
         category,
         city,
@@ -276,15 +315,14 @@ export const PoiController: IController = {
           res
             .status(400)
             .send({ error: `Field must contains only characters` });
-          await unlink(`./public/poi/${req.file?.filename}`);
-
-          return;
+          if (req.file !== undefined)
+            await unlink(`./public/poi/${req.file.filename}`);
         }
       };
 
       const inputString: string[] = [description, name, address];
-      inputString.forEach((value) => {
-        if (value) checkIfStringAndNotEmpty(value);
+      inputString.forEach(async (value) => {
+        if (value !== undefined) await checkIfStringAndNotEmpty(value);
       });
 
       // check coordinates format
@@ -292,14 +330,15 @@ export const PoiController: IController = {
         res.status(400).send({
           error: "Incorrect format of coordinates (must be [lat, long])",
         });
-        await unlink(`./public/poi/${req.file?.filename}`);
+        if (req.file !== undefined)
+          await unlink(`./public/poi/${req.file.filename}`);
 
         return;
       }
 
       // check name
       if (
-        name &&
+        name !== null &&
         !validator.matches(
           name,
           /^[a-zA-ZàáâäãåąčćęèéêëėįìíîïłńòóôöõøùúûüųūÿýżźñçčšžÀÁÂÄÃÅĄĆČĖĘÈÉÊËÌÍÎÏĮŁŃÒÓÔÖÕØÙÚÛÜŲŪŸÝŻŹÑßÇŒÆČŠŽ∂ð '-]{2,100}$/
@@ -308,15 +347,15 @@ export const PoiController: IController = {
         res.status(400).send({
           error: `Field must contains only characters (min: 2, max: 100)`,
         });
-        await unlink(`./public/poi/${req.file?.filename}`);
-
+        if (req.file !== undefined)
+          await unlink(`./public/poi/${req.file.filename}`);
         return;
       }
 
       // check address
 
       if (
-        address &&
+        address !== null &&
         !validator.matches(
           address,
           /^(([\u00c0-\u01ffa-zA-Z\d])+(?:['-\s][\u00c0-\u01ffa-zA-Z]+)*\s)+((\d{5})*\s)([\u00c0-\u01ffa-zA-Z])+(?:['-\s][\u00c0-\u01ffa-zA-Z]+)*\s*$/
@@ -325,7 +364,8 @@ export const PoiController: IController = {
         res.status(400).send({
           error: `Incorrect format of address`,
         });
-        await unlink(`./public/poi/${req.file?.filename}`);
+        if (req.file !== undefined)
+          await unlink(`./public/poi/${req.file.filename}`);
 
         return;
       }
@@ -337,22 +377,22 @@ export const PoiController: IController = {
           res.status(400).send({
             error: "Incorrect format of foreign key (must be uuid)",
           });
-          await unlink(`./public/poi/${req.file?.filename}`);
-
-          return;
+          if (req.file !== undefined)
+            await unlink(`./public/poi/${req.file?.filename}`);
         }
       };
 
       const foreignKeys: string[] = [category, city, user];
-      foreignKeys.forEach((value) => {
-        if (value) checkIfUUID(value);
+      foreignKeys.forEach(async (value) => {
+        if (value !== undefined) await checkIfUUID(value);
       });
 
       // check if is_accepted is boolean
 
-      if (is_accepted && typeof is_accepted !== "boolean") {
+      if (isAccepted !== null && typeof isAccepted !== "boolean") {
         res.status(400).send({ error: "is_accepted must be a boolean" });
-        await unlink(`./public/poi/${req.file?.filename}`);
+        if (req.file !== undefined)
+          await unlink(`./public/poi/${req.file.filename}`);
         return;
       }
 
@@ -360,8 +400,8 @@ export const PoiController: IController = {
       const poiToUpdate = await dataSource.getRepository(Poi).findOneBy({ id });
       if (poiToUpdate === null) {
         res.status(404).send({ error: "Point of interest not found" });
-        await unlink(`./public/poi/${req.file?.filename}`);
-
+        if (req.file !== undefined)
+          await unlink(`./public/poi/${req.file.filename}`);
         return;
       }
 
@@ -386,8 +426,8 @@ export const PoiController: IController = {
         res.status(403).send({
           error: "You are not authorized to create a point of interest",
         });
-        await unlink(`./public/poi/${req.file?.filename}`);
-
+        if (req.file !== undefined)
+          await unlink(`./public/poi/${req.file.filename}`);
         return;
       }
 
@@ -404,8 +444,8 @@ export const PoiController: IController = {
         });
         if (coordsAlreadyExist > 0) {
           res.status(404).send({ error: "Point of interest already exist" });
-          await unlink(`./public/poi/${req.file?.filename}`);
-
+          if (req.file !== undefined)
+            await unlink(`./public/poi/${req.file.filename}`);
           return;
         }
 
@@ -417,7 +457,7 @@ export const PoiController: IController = {
       }
 
       // rename the new file and delete the older one
-      if (req.file) {
+      if (req.file !== undefined) {
         // rename
         const originalname = req.file.originalname;
         const filename = req.file.filename;
@@ -426,7 +466,7 @@ export const PoiController: IController = {
           `./public/poi/${filename}`,
           `./public/poi/${newName}`,
           (err) => {
-            if (err) throw err;
+            if (err !== null) throw err;
           }
         );
         req.body.image = `/public/poi/${newName}`;
@@ -443,7 +483,8 @@ export const PoiController: IController = {
       res.status(200).send("Updated point of interest");
     } catch (err) {
       res.status(400).send({ error: "Error while updating point of interest" });
-      await unlink(`./public/poi/${req.file?.filename}`);
+      if (req.file !== undefined)
+        await unlink(`./public/poi/${req.file.filename}`);
     }
   },
 
